feat: resolve task2
