+++ bohye/board-project/src/components/week9/task2/App.js
@@ -1,13 +1,8 @@
-import { useState, useEffect } from "react";
+import { useCounter } from "./useCounter.js";
 import "./styles.css";
 
 export default function Counter() {
-  const [count, setCount] = useState(0);
-  useEffect(() => {
-    const id = setInterval(() => {
-      setCount((c) => c + 1);
-    }, 1000);
-    return () => clearInterval(id);
-  }, []);
+  const count = useCounter();
+
   return <h1>Seconds passed: {count}</h1>;
 }
