+++ bohye/board-project/src/components/week9/task2/useCounter.js
@@ -1,0 +1,14 @@
+import { useState, useEffect } from "react";
+
+export function useCounter() {
+  const [count, setCount] = useState(0);
+
+  useEffect(() => {
+    const id = setInterval(() => {
+      setCount((c) => c + 1);
+    }, 1000);
+    return () => clearInterval(id);
+  }, []);
+
+  return count;
+}
